feat: add tslint / fixs bugs
