+++ src/datastructure/queue/index.ts
@@ -1,5 +1,5 @@
-import Queue from './Queue'
-import Dequeue from './Dequeue'
+import Queue from '@ds/Queue/Queue'
+import Dequeue from '@ds/Queue/Dequeue'
 
 export { Queue, Dequeue }
 export default Queue
