+++ src/datastructure/queue/queue.test.ts
@@ -1,4 +1,4 @@
-import Queue from './index'
+import Queue from '@ds/Queue'
 
 test('test Queue ', () => {
   const q = new Queue()
