+++ tslint.json
@@ -1,0 +1,10 @@
+{
+    "defaultSeverity": "error",
+    "extends": ["tslint-microsoft-contrib/recommended"],
+    "rulesDirectory": ["node_modules/tslint-microsoft-contrib"],
+    "rules":[
+       {
+           "perfer-cast":false
+       }
+    ]
+}
